refactor: add-search
