+++ src/main/java/ru/yandex/practicum/filmorate/storage/director/DirectorDbStorage.java
@@ -9,7 +9,6 @@ import ru.yandex.practicum.filmorate.exception.EntityNotFoundException;
 import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.Validator;
-import ru.yandex.practicum.filmorate.storage.film.FilmDbStorage;
 import ru.yandex.practicum.filmorate.utilities.Checker;
 
 import java.sql.PreparedStatement;
@@ -81,5 +80,16 @@ public class DirectorDbStorage implements DirectorDaoStorage {
     public void updateDirectorFilm(Film film) {
         String sql = "DELETE FROM FILM_DIRECTOR WHERE FILM_ID = ?";
         jdbcTemplate.update(sql, film.getId());
+    }
+
+    @Override
+    public Director updateDirector(Director director) {
+        Checker.checkDirectorExists(director.getId(), jdbcTemplate);
+        jdbcTemplate.update("UPDATE DIRECTORS SET DIRECTOR_NAME = ? WHERE DIRECTOR_ID = ?"
+                , director.getName()
+                , director.getId());
+        jdbcTemplate.update("UPDATE FILM_DIRECTOR set DIRECTOR_ID = ? WHERE FILM_ID",
+                director.getId());
+        return director;
     }
 }
