+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDaoStorage.java
@@ -44,6 +44,5 @@ public interface FilmDaoStorage {
 
     List<Film> getSearchFilmsForTitleAndDirector(String query);
 
-    Director updateDirector(Director director);
 }
