+++ src/main/java/ru/yandex/practicum/filmorate/storage/like/LikeDaoStorage.java
@@ -1,7 +1,5 @@
 package ru.yandex.practicum.filmorate.storage.like;
 
-import ru.yandex.practicum.filmorate.model.Film;
-
 public interface LikeDaoStorage {
 
     void saveLikes(Long id, Long userId);
