+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaDaoStorage.java
@@ -1,6 +1,5 @@
 package ru.yandex.practicum.filmorate.storage.mpa;
 
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Mpa;
 
 import java.util.List;
